사용자 검증 추가, 반복문 스트림으로 변경

+++ src/main/java/com/coolpeace/domain/accommodation/repository/AccommodationRepository.java
@@ -9,7 +9,7 @@ public interface AccommodationRepository extends JpaRepository<Accommodation, Lo
 
     List<Accommodation> findAllByMember(Member member);
 
-    List<Accommodation> findAllByMemberId(Integer memberId);
+    List<Accommodation> findAllByMemberId(Long memberId);
 
     List<Accommodation> findByAddress(String address);
 }
